Add RicardAgrawala algorithm for mutual exclusion

+++ src/main/kotlin/wjaronski/message/LogicalClock.kt
@@ -1,22 +1,22 @@
 package wjaronski.message
 
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 class LogicalClock {
-    private val value = AtomicInteger(0)
+    private val value = AtomicLong(0)
 
     @Synchronized
-    fun current(): Int {
+    fun current(): Long {
         return value.get()
     }
 
     @Synchronized
-    fun next(): Int {
+    fun next(): Long {
         return value.incrementAndGet()
     }
 
     @Synchronized
-    fun sync(value: Int): Int {
+    fun sync(value: Long): Long {
         value.coerceAtLeast(value)
         return next()
     }
